рефакторинг редьюсера
