+++ src/components/app/app.test.js
@@ -4,14 +4,19 @@ import {Provider} from "react-redux";
 import configureStore from "redux-mock-store";
 import App from "./app.jsx";
 import {mockFilmCard, mockFilms, mokcFunction, MOCK_FILMS_COUNT, mockBool} from "../../mocks-for-tests.js";
+import Namespace from "../../reducer/namespace.js";
 
 const mockStore = configureStore([]);
 
 describe(`Render correct App`, () => {
   it(`Render App`, () => {
     const store = mockStore({
-      films: mockFilms,
-      filmsCount: MOCK_FILMS_COUNT
+      [Namespace.DATA]: {
+        films: mockFilms,
+      },
+      [Namespace.STATE]: {
+        filmsCount: MOCK_FILMS_COUNT
+      }
     });
     const tree = renderer
     .create(
