+++ src/components/player/player.jsx
@@ -27,7 +27,7 @@ Player.propTypes = {
     votes: PropTypes.number,
     director: PropTypes.string,
     starring: PropTypes.arrayOf(PropTypes.string),
-    runTime: PropTypes.number,
+    runTime: PropTypes.string,
     genre: PropTypes.string,
     releaseYear: PropTypes.number,
     id: PropTypes.number,
