+++ src/index.js
@@ -2,12 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom";
 import {createStore, applyMiddleware} from "redux";
 import {Provider} from "react-redux";
-import {reducer, Operation as DataOperation} from "./reducer/reducer.js";
+import reducer from "./reducer/reducer.js";
 import thunk from "redux-thunk";
 import App from "./components/app/app.jsx";
 // import films from "./mocks/films.js";
 import FilmCard from "./mocks/film-card.js";
 import {createAPI} from "./api.js";
+import {Operation as DataOperation} from "./reducer/data/data.js";
 
 const api = createAPI();
