+++ src/reducer/data/data.test.js
@@ -1,22 +1,21 @@
-import {reducer, ActionType, ActionCreator, Operation} from "./reducer.js";
+import {reducer, ActionType, ActionCreator, Operation} from "../../reducer/data/data.js";
 import MockAdapter from "axios-mock-adapter";
-import {createAPI} from "../api.js";
-import {mockFilms} from "../mocks-for-tests.js";
+import {createAPI} from "../../api.js";
+import {mockFilms} from "../../mocks-for-tests.js";
 
 const api = createAPI(() => {});
 
-it(`Reducer without additional parameters should return initial state`, () => {
+it(`Reducer without additional parameters should return initial state for films and active filter`, () => {
   expect(reducer(void 0, {})).toEqual({
     films: [],
-    filmsCount: 8,
     activeFilter: `All genres`
   });
 });
 
-describe(`Reducer return current value`, () => {
+describe(`Reducer return current value for films and filters`, () => {
   it(`Reducer should change current film list by genre`, () => {
     expect(reducer({
-      ilms: mockFilms,
+      films: mockFilms,
       activeFilter: `Drama`
     }, {
       type: ActionType.CHANGE_GENRE,
@@ -47,45 +46,13 @@ describe(`Reducer return current value`, () => {
     });
   });
 
-  it(`Reducer should change films count`, () => {
-    expect(reducer({
-      filmsCount: 8
-    }, {
-      type: ActionType.SHOW_MORE_FILMS,
-      payload: null,
-    })).toEqual({
-      filmsCount: 8 + 8
-    });
-    expect(reducer({
-      filmsCount: 16
-    }, {
-      type: ActionType.SHOW_MORE_FILMS,
-      payload: null,
-    })).toEqual({
-      filmsCount: 16 + 8
-    });
-    expect(reducer({
-      filmsCount: 24
-    }, {
-      type: ActionType.SHOW_MORE_FILMS,
-      payload: null,
-    })).toEqual({
-      filmsCount: 24 + 8
-    });
-  });
 });
 
-describe(`Action creators work correctly`, () => {
+describe(`Action creators work correctly for action CHANGE GENRE`, () => {
   it(`Action creator for change genre returns correct action`, () => {
     expect(ActionCreator.changeGenre(`All genres`)).toEqual({
       type: ActionType.CHANGE_GENRE,
       payload: `All genres`
-    });
-  });
-  it(`Action creator for show more films returns correct action`, () => {
-    expect(ActionCreator.showMoreFilms()).toEqual({
-      type: ActionType.SHOW_MORE_FILMS,
-      payload: null
     });
   });
 });
@@ -110,5 +77,3 @@ describe(`Operation work correctly`, () => {
       });
   });
 });
-
-
