+++ src/reducer/state/selectors.js
@@ -1,0 +1,5 @@
+import Namespace from "../namespace.js";
+
+export const getShowMoreFilms = (state) => {
+  return state[Namespace.STATE].filmsCount;
+};
